+++ core/tauri/src/lib.rs
@@ -989,26 +989,6 @@ impl From<IconDto> for Icon {
 }
 
 #[allow(unused)]
-macro_rules! run_item_main_thread {
-  ($self:ident, $ex:expr) => {{
-    use std::sync::mpsc::channel;
-    let (tx, rx) = channel();
-    let self_ = $self.clone();
-    let task = move || {
-      let f = $ex;
-      let _ = tx.send(f(self_));
-    };
-    $self
-      .app_handle()
-      .run_on_main_thread(task)
-      .and_then(|_| rx.recv().map_err(|_| crate::Error::FailedToReceiveMessage))
-  }};
-}
-
-#[allow(unused)]
-pub(crate) use run_item_main_thread;
-
-#[allow(unused)]
 macro_rules! run_main_thread {
   ($handle:ident, $ex:expr) => {{
     use std::sync::mpsc::channel;
