+++ core/tauri/src/menu/check.rs
@@ -4,9 +4,10 @@
 
 use std::sync::Arc;
 
+use super::run_item_main_thread;
 use crate::menu::CheckMenuItemInner;
+use crate::run_main_thread;
 use crate::{menu::MenuId, AppHandle, Manager, Runtime};
-use crate::{run_item_main_thread, run_main_thread};
 
 use super::CheckMenuItem;
