+++ core/tauri/src/menu/icon.rs
@@ -4,10 +4,11 @@
 
 use std::sync::Arc;
 
+use super::run_item_main_thread;
 use super::{IconMenuItem, NativeIcon};
 use crate::menu::IconMenuItemInner;
+use crate::run_main_thread;
 use crate::{menu::MenuId, AppHandle, Icon, Manager, Runtime};
-use crate::{run_item_main_thread, run_main_thread};
 
 impl<R: Runtime> IconMenuItem<R> {
   /// Create a new menu item.
