+++ core/tauri/src/menu/menu.rs
@@ -4,12 +4,13 @@
 
 use std::sync::Arc;
 
+use super::run_item_main_thread;
 use super::sealed::ContextMenuBase;
 use super::{
   AboutMetadata, IsMenuItem, Menu, MenuInner, MenuItemKind, PredefinedMenuItem, Submenu,
 };
+use crate::run_main_thread;
 use crate::Window;
-use crate::{run_item_main_thread, run_main_thread};
 use crate::{AppHandle, Manager, Position, Runtime};
 use muda::ContextMenu;
 use muda::MenuId;
