+++ core/tauri/src/menu/predefined.rs
@@ -4,10 +4,11 @@
 
 use std::sync::Arc;
 
+use super::run_item_main_thread;
 use super::{AboutMetadata, PredefinedMenuItem};
 use crate::menu::PredefinedMenuItemInner;
+use crate::run_main_thread;
 use crate::{menu::MenuId, AppHandle, Manager, Runtime};
-use crate::{run_item_main_thread, run_main_thread};
 
 impl<R: Runtime> PredefinedMenuItem<R> {
   /// Separator menu item
